Continue on the Distributor/Scheduler Layout
Running the test.py on test/qstack/distributor gives me an error
I cannot find why: "'Scheduler' object has no attribute 'queue'".
It should work.

Changes:
- The Distributor does not need a main it just assigns the Qernels
to the backends as they arrive (register_qernel)

- We might have a small circular import (types<-->scheduler) that it
is now solved but I had to create an abstract class for the scheduler
but it is pretty much useless, it just solves the circular import